created package 'antiweb' with module directives
